✨ [12] 인증 요청

+++ TodoApp/main.py
@@ -6,6 +6,7 @@ from starlette import status
 from pydantic import BaseModel, Field
 
 import models
+from auth import get_current_user, User, get_user_exception
 from database import engine, SessionLocal
 from sqlalchemy.orm import Session
 
@@ -34,12 +35,23 @@ async def read_all(db: Session = Depends(get_db)):
     return db.query(models.Todo).all()
 
 
+@app.get("/todos/user")
+async def read_all_by_user(user: User = Depends(get_current_user), db: Session = Depends(get_db)):
+    if user is None:
+        raise get_user_exception()
+    return db.query(models.Todo).filter(models.Todo.owner_id == user.id).all()
+
+
 @app.post("/", status_code=status.HTTP_201_CREATED)
-async def create_todo(todo: Todo, db: Session = Depends(get_db)):
+async def create_todo(todo: Todo, user: User = Depends(get_current_user), db: Session = Depends(get_db)):
+    if user is None:
+        raise get_user_exception()
+
     new_todo = models.Todo(title=todo.title,
                            description=todo.description,
                            priority=todo.priority,
-                           complete=todo.complete)
+                           complete=todo.complete,
+                           owner_id=user.id)
     db.add(new_todo)
     db.commit()
     db.refresh(new_todo)
@@ -47,17 +59,26 @@ async def create_todo(todo: Todo, db: Session = Depends(get_db)):
     return new_todo
 
 
-@app.get("/{todo_id}")
-async def read_todo(todo_id: int, db: Session = Depends(get_db)):
-    todo = db.query(models.Todo).filter(models.Todo.id == todo_id).first()
+@app.get("/todo/{todo_id}")
+async def read_todo(todo_id: int, user: User = Depends(get_current_user), db: Session = Depends(get_db)):
+    todo = db.query(models.Todo) \
+        .filter(models.Todo.id == todo_id) \
+        .filter(models.Todo.id == user.id) \
+        .first()
     if todo is None:
         raise_http_exception()
     return todo
 
 
-@app.put("/{todo_id}/", status_code=status.HTTP_202_ACCEPTED)
-async def update_todo(todo_id: int, todo: Todo, db: Session = Depends(get_db)):
-    todo_model = db.query(models.Todo).filter(models.Todo.id == todo_id).first()
+@app.put("/todo/{todo_id}/", status_code=status.HTTP_202_ACCEPTED)
+async def update_todo(todo_id: int, todo: Todo, user: User = Depends(get_current_user), db: Session = Depends(get_db)):
+    if user is None:
+        raise get_user_exception()
+
+    todo_model = db.query(models.Todo) \
+        .filter(models.Todo.id == todo_id) \
+        .filter(models.Todo.owner_id == user.id) \
+        .first()
     if todo is None:
         raise_http_exception()
     todo_model.title = todo.title
@@ -72,9 +93,15 @@ async def update_todo(todo_id: int, todo: Todo, db: Session = Depends(get_db)):
     return todo_model
 
 
-@app.delete("/{todo_id}")
-async def delete_todo(todo_id: int, db: Session = Depends(get_db)):
-    todo_model = db.query(models.Todo).filter(models.Todo.id == todo_id).first()
+@app.delete("/todo/{todo_id}")
+async def delete_todo(todo_id: int, user: User = Depends(get_current_user), db: Session = Depends(get_db)):
+    if user is None:
+        raise get_user_exception()
+
+    todo_model = db.query(models.Todo) \
+        .filter(models.Todo.id == todo_id) \
+        .filter(models.Todo.owner_id == user.id) \
+        .first()
     if todo_model is None:
         raise_http_exception()
     db.query(models.Todo).filter(models.Todo.id == todo_id).delete()
